fix: integer rounding for CEA/EUA, commission flush bug, bytes serialization + role flow features
- Fix commission_ledger IntegrityError: add db.flush() after db.add(trade) in
  order_matching.py so trade.id is populated before maybe_create_commission
- Fix ticket_service.get_entity_state: handle bytes columns (nda_file_data)
  to prevent JSON serialization error on deposit reporting
- CEA/EUA quantities floored to integer at settlement creation (in favor of NIHA):
  order_matching.py uses int(preview.total_quantity) for CEA purchase settlements;
  swaps.py uses int(eua_needed * (1 - fee)) for EUA swap settlements
- Add PRE_NDA role flow: contact.py, admin.py, email templates, frontend pages
- Add KycFormWizard component, IntroducerPage flow, SetupPasswordPage updates
- Add alembic migrations: volume caps, market settings, entity_type, kyc_form_data,
  cea_balance_mm, sync_mm_balances
- Add pdf_generator service (WeasyPrint): custody, derivatives, fee_schedule,
  kyc, msa, risk_disclosure generators
- Add NIHA business documents and design files

+++ backend/app/api/v1/swaps.py
@@ -930,9 +930,9 @@ async def execute_swap(
             detail=f"Insufficient swap liquidity. Can only swap {cea_quantity - remaining_cea:.2f} CEA out of {cea_quantity:.2f} requested.",
         )
 
-    # Apply platform fee (0.5%)
+    # Apply platform fee (0.5%) and floor to integer — fractional EUA stays with NIHA
     platform_fee_pct = 0.005
-    eua_output = eua_needed * (1 - platform_fee_pct)
+    eua_output = int(eua_needed * (1 - platform_fee_pct))
 
     # Calculate weighted average ratio
     total_cea_matched = sum(m["cea_amount"] for m in matched_orders)
